피드 상세 페이지

+++ src/main/java/com/project/sangil_be/repository/GoodRepository.java
@@ -16,7 +16,7 @@ public interface GoodRepository extends JpaRepository<Good, Long> {
 
     List<Good> findByFeedId(Long feed);
 
-    Long countAllByFeedId(Long feedId);
+    int countAllByFeedId(Long feedId);
 
     Long countAllByUserId(Long userId);
 }
